requirements.txt
django
factory-boy
psycopg2-binary